Finishing up export workflow, and starting to add Result to some table list endpoints
