+++ mason/clients/glue/metastore.py
@@ -6,12 +6,12 @@ from mason.clients.engines.metastore import MetastoreClient
 from mason.clients.engines.valid_client import ValidClient
 from mason.clients.glue.glue_client import GlueClient
 from mason.clients.response import Response
-from mason.engines.execution.models.jobs import ExecutedJob, InvalidJob, Job
+from mason.engines.execution.models.jobs import ExecutedJob, InvalidJob
 from mason.engines.metastore.models.credentials import InvalidCredentials
 from mason.engines.metastore.models.credentials.aws import AWSCredentials
 from mason.engines.metastore.models.database import Database, InvalidDatabase
 from mason.engines.metastore.models.ddl import DDLStatement, InvalidDDLStatement
-from mason.engines.metastore.models.table import Table, InvalidTables
+from mason.engines.metastore.models.table import Table, InvalidTables, TableList
 from mason.engines.storage.models.path import Path
 
 class GlueMetastoreClient(MetastoreClient, ValidClient):
@@ -25,7 +25,7 @@ class GlueMetastoreClient(MetastoreClient, ValidClient):
     def delete_table(self, database_name: str, table_name: str, response: Optional[Response] = None) -> Response:
         return self.client.delete_table(database_name, table_name, response)
 
-    def list_tables(self, database_name: str, response: Response) -> Tuple[Result[List[Table], InvalidTables], Response]:
+    def list_tables(self, database_name: str, response: Response) -> Tuple[Result[TableList, InvalidTables], Response]:
         return self.client.list_tables(database_name, response)
 
     def get_table(self, database_name: str, table_name: str, options: Optional[dict] = None, response: Optional[Response] = None) -> Tuple[Union[Table, InvalidTables], Response]:
