+++ mason/clients/s3/metastore.py
@@ -1,5 +1,5 @@
 from urllib.parse import urlparse
-from typing import Tuple, Union, Optional
+from typing import Tuple, Union, Optional, List
 
 from returns.result import Result
 
@@ -7,14 +7,13 @@ from mason.clients.engines.metastore import MetastoreClient
 from mason.clients.engines.valid_client import ValidClient
 from mason.clients.response import Response
 from mason.clients.s3.s3_client import S3Client
-from mason.engines.execution.models.jobs import ExecutedJob, InvalidJob, Job
+from mason.engines.execution.models.jobs import ExecutedJob, InvalidJob
 from mason.engines.metastore.models.credentials import InvalidCredentials
 from mason.engines.metastore.models.credentials.aws import AWSCredentials
 from mason.engines.metastore.models.database import Database, InvalidDatabase
 from mason.engines.metastore.models.ddl import DDLStatement, InvalidDDLStatement
-from mason.engines.metastore.models.table import Table, InvalidTables
+from mason.engines.metastore.models.table import Table, InvalidTables, TableList
 from mason.engines.storage.models.path import Path
-
 
 class S3MetastoreClient(MetastoreClient, ValidClient):
 
@@ -26,7 +25,7 @@ class S3MetastoreClient(MetastoreClient, ValidClient):
         database = tables.map(lambda a: Database("s3_table", a)).alt(lambda b: InvalidDatabase(b.error or b.message()))
         return database, response
 
-    def list_tables(self, database_name: str, response: Response) -> Tuple[Result[List[Table], InvalidTables], Response]:
+    def list_tables(self, database_name: str, response: Response) -> Tuple[Result[TableList, InvalidTables], Response]:
         return self.client.list_tables(database_name, response)
 
     def get_table(self, database_name: str, table_name: str, options: Optional[dict] = None, response: Optional[Response] = None) -> Tuple[Union[Table, InvalidTables], Response]:
