+++ mason/clients/s3/s3_client.py
@@ -6,7 +6,7 @@ from s3fs import S3FileSystem
 
 from mason.clients.aws_client import AWSClient
 from mason.clients.response import Response
-from mason.engines.metastore.models.table import InvalidTable, Table, TableNotFound, InvalidTables
+from mason.engines.metastore.models.table import InvalidTable, Table, TableNotFound, InvalidTables, TableList
 from mason.engines.storage.models.path import Path
 from mason.util.exception import message
 from mason.util.logger import logger
@@ -19,7 +19,6 @@ class S3Client(AWSClient):
     def __init__(self, s3_config: dict):
         super().__init__(**s3_config)
         
-
     def client(self) -> S3FileSystem:
         s3 = s3fs.S3FileSystem(key=self.access_key, secret=self.secret_key, client_kwargs={'region_name': self.aws_region})
 
@@ -54,15 +53,27 @@ class S3Client(AWSClient):
     def parse_items(self, s3_response: dict):
         return list(map(lambda x: self.parse_item(x), s3_response.get('Contents', [])))
 
-    @safe
-    def list_objects(self, database_name: str) -> dict:
-        split = database_name.split("/", 1)
-        return self.client().s3.list_objects(Bucket=split[0], Prefix=(get(split, 1) or '/'), Delimiter='/')
+    def list_objects(self, database_name: str, response: Response) -> Tuple[Result[dict, str], Response]:
+        try:
+            split = database_name.split("/", 1)
+            result = self.client().s3.list_objects(Bucket=split[0], Prefix=(get(split, 1) or '/'), Delimiter='/')
+            response.add_response(result)
+            return Success(result), response
+        except Exception as e:
+            if isinstance(e, ClientError):
+                result = e.response
+                error = result.get("Error", {})
+                code = error.get("Code", "")
+                if code == "NoSuchBucket":
+                    response.set_status(404)
+                    return Failure(f"The specified bucket does not exist: {database_name}"), response
+            return Failure(message(e)), response
 
-    def list_tables(self, database_name: str, response: Response) -> Tuple[Result[List[Table], InvalidTables], Response]:
-        result: Result[dict, InvalidTables] = self.list_objects(database_name).alt(lambda e: InvalidTables([], message(e)))
+    def list_tables(self, database_name: str, response: Response) -> Tuple[Result[TableList, InvalidTables], Response]:
+        tables, response = self.list_objects(database_name, response)
+        result: Result[dict, InvalidTables] = tables.alt(lambda e: InvalidTables([], e))
         
-        def parse_response(result: dict, response: Response) -> Result[List[Table], InvalidTables]:
+        def parse_response(result: dict, response: Response) -> Result[TableList, InvalidTables]:
             contents: Optional[List[dict]] = result.get("Contents")
             prefixes: Optional[List[dict]] = result.get("CommonPrefixes")
             
@@ -75,7 +86,7 @@ class S3Client(AWSClient):
                         tables.append(table)
                 valid, invalid = sequence(tables, Table, InvalidTables) 
                 if len(valid) > 0:
-                    return Success(valid)
+                    return Success(TableList(valid))
                 else:
                     invalid_tables: List[InvalidTable] = []
                     for i in invalid:
@@ -89,20 +100,9 @@ class S3Client(AWSClient):
             else:
                 return Failure(InvalidTables([], "No Data returned from AWS"))
 
+        # TODO:  response is not pure here
         final = result.bind(lambda r: parse_response(r, response))
-        
-        # except ClientError as e:
-        #     result = e.response
-        #     error = result.get("Error", {})
-        #     code = error.get("Code", "")
-        #     if code == "NoSuchBucket":
-        #         response.add_error(error.get("Message") + f": {database_name}")
-        #         response.set_status(404)
-        #     else:
-        #         raise e
-        # 
-        # response.add_response(result)
-        
+
         return final, response
 
     def get_table(self, database_name: str, table_name: str, options: Optional[dict] = None, response: Optional[Response] = None) -> Tuple[Union[Table, InvalidTables], Response]:
