+++ mason/engines/metastore/models/database/__init__.py
@@ -1,10 +1,8 @@
-from typing import List
-
-from mason.engines.metastore.models.table import Table
+from mason.engines.metastore.models.table import TableList
 
 class Database:
 
-    def __init__(self, name: str, tables: List[Table]):
+    def __init__(self, name: str, tables: TableList):
         self.name = name
         self.tables = tables
