+++ mason/examples/operators/table/test/expects/table.py
@@ -17,8 +17,8 @@ def index(client_type: str, exists=True):
             return ({'Errors': ['Database bad-database not found'], 'Info': [], 'Warnings': []}, 404)
     elif client_type == "s3":
         if exists:
-            return ({'Data': [{'Prefixes': [{'Prefix': '/test_path/'}]}],
-              'Errors': [],
+            return ({'Data': [{'Prefix': '/test_path/'}],
+              'Errors': ["No valid tables at crawler-poc.  Try appending '/' or specify deeper key."],
               'Info': [],
               'Warnings': []},
              200)
