+++ requirements.txt
@@ -31,7 +31,7 @@ cloudpickle==1.5.0
 connexion>=2.6
 swagger-ui-bundle
 connexion[swagger-ui]>2.6
-mason_dask==0.1.22
+mason_dask==0.1.24
 Markdown==3.2.2
 lambdas==0.1.0
 returns==0.14.0
